refactor(cli): colocate tests with the modules they test
Each module's tests now sit beside it — args.test.ts, run-input.test.ts,
main.test.ts and one file per src/commands/* — instead of a package-level
commands.test.ts. The shared harness moves to src/test-harness.ts. Only
bin.test.ts stays under test/: it exercises bin/rcv.mjs, which is not a
src module, alongside the fixtures both suites share.
