+++ packages/cli/vitest.config.ts
@@ -9,6 +9,11 @@ import { defineConfig } from "vitest/config";
  * host for the runner. Thin by design: the golden↔shimmed parity suite in
  * `packages/engine` is what proves the resolution semantics; these tests cover
  * argument parsing, output shapes and exit codes.
+ *
+ * A test sits next to the module it covers, so the suites live under `src/`.
+ * The exception is `test/bin.test.ts`: `bin/rcv.mjs` is not a `src` module, and
+ * `bin/` is published payload that carries no tests. `test/fixtures/` are the
+ * configs both regimes feed in.
  */
 export default defineConfig({
   resolve: {
@@ -19,7 +24,7 @@ export default defineConfig({
   plugins: [renovateShims()],
   test: {
     name: "cli",
-    include: ["test/*.test.ts"],
+    include: ["src/**/*.test.ts", "test/*.test.ts"],
     environment: "node",
     // the first test to resolve a large internal preset pays the lazy
     // transform+import of renovate's preset data modules
